fix : search function

+++ core/network/src/main/java/com/yudistiro/network/model/WeatherInfo.kt
@@ -10,7 +10,7 @@ data class WeatherInfo (
   @SerializedName("weather" ) var weather : ArrayList<Weather> = arrayListOf(),
   @SerializedName("clouds"  ) var clouds  : Clouds?            = Clouds(),
   @SerializedName("wind"    ) var wind    : Wind?              = Wind(),
-  @SerializedName("pop"     ) var pop     : Int?               = null,
+  @SerializedName("pop"     ) var pop     : Double?               = null,
   @SerializedName("sys"     ) var sys     : Sys?               = Sys(),
   @SerializedName("dt_txt"  ) var dtTxt   : String?            = null
